requirements.txt
flask
flask-cors
skyfield
numpy
pytz
pyswisseph
